fix for ubsan on unittest.h fastrand()
internal math of the fastrand function uses a multiply
and add that overflows a signed int.  This triggers a
ubsan failure:

../../unit_test/../unit_test/unit_test.h:60:33: runtime error: signed integer overflow: 56248274 * 214013 cannot be represented in type 'int'

This change casts the intermediate math to unsigned
int to avoid the overflow.

For more info on ubsan, see
http://dev.chromium.org/developers/testing/undefinedbehaviorsanitizer

TESTED=Passing compilation using:
GYP_DEFINES="ubsan=1"
GYP_DEFINES="ubsan_vptr=1"

R=harryjin@google.com, pbos@webrtc.org
BUG=libyuv:563

Review URL: https://codereview.chromium.org/1662453003 .

+++ unit_test/unit_test.h
@@ -55,10 +55,10 @@ static inline double get_time() {
 }
 #endif
 
-extern int fastrand_seed;
+extern unsigned int fastrand_seed;
 inline int fastrand() {
-  fastrand_seed = fastrand_seed * 214013 + 2531011;
-  return (fastrand_seed >> 16) & 0xffff;
+  fastrand_seed = fastrand_seed * 214013u + 2531011u;
+  return static_cast<int>((fastrand_seed >> 16) & 0xffff);
 }
 
 static inline void MemRandomize(uint8* dst, int64 len) {
